// types/categories.ts
export enum Categories {
  BUILD = 'build',
  CLOUD = 'cloud',
  DATABASES = 'databases',
  FRAMEWORK = 'framework',
  INTERESTS = 'interests',
  LIBRARY = 'library',
  OTHER = 'other',
  PROGRAMMING_LANGUAGE = 'programming-language',
  SAAS_PAAS = 'saas-paas',
  SERVER = 'server',
  WANT_TO_LEARN = 'want-to-learn',
}